Add transformers support to transient
Transient is now a transformer monad named AsyncT that provides
composition of asynchronous effects on top of any monad that supports
MonadIO. It can be used anywhere in a transformer stack.

TransIO is now implemented as "AsyncT IO".  There is no change in
external APIs except that now you have to use a new setSData API in place
of setData inside TransIO, just like getSData for get operation.

The change consists of largely signature changes. I have tried to
keep the changes to a minimum in this commit to help in review. The
significant changes are:

1) Add the base monad as an existential parameter to EventF
2) A few more unsafeCoerces for the continuation type conversions
3) lowering of monad computations into IO to facilitate forkIO for new
   event creation. This requires monad-control and lifted-base.